Docs: Finalized (and tested) enum docs
