+++ docs/enum.md
@@ -1,6 +1,1 @@
 ::: py_back.enum
-
-> Backported from version 3.11
-
-_ReprEnum_ uses the repr() of Enum, but the str() of the mixed-in data type.
-The class is used for any builtin type enum.
