+++ src/py_back/enum/__init__.py
@@ -1,4 +1,24 @@
-"""Module to backport 'enum' classes depending on the system python."""
+"""Backported enum types.
+
+All enumerations can be imported just like the ones provided by python.
+```python
+from py_back import enum
+
+class Number(enum.IntEnum):
+    \"\"\"Enumeration using the original 'IntEnum' call\"\"\"
+    ONE = enum.auto()
+    TWO = 2
+
+
+class Animal(enum.StrEnum):
+    \"\"\"Supported original 'StrEnum' for python versions < 3.11\"\"\"
+    CAT = enum.auto()
+    DOG = "dog"
+```
+
+Importing the `py_back.enum` module ensures that only the required classes are
+backported.
+"""
 
 import sys
 
@@ -15,9 +35,7 @@ __all__ = [
     "unique",
 ]
 
-if sys.version_info < (3, 11):
-    from enum import Enum, EnumMeta, Flag, auto, unique
+from enum import *
 
+if sys.version_info < (3, 11):
     from ._back_enum import EnumType, IntEnum, IntFlag, ReprEnum, StrEnum
-else:
-    from enum import *
